# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentación de mi API Rest\n\nPara autenticarte:\n1. Usa el endpoint /auth/login para obtener el token\n2. Copia el token devuelto\n3. Click en el botón 'Authorize' (🔓) arriba\n4. Pega el token en el campo 'Value' (incluye 'Bearer ')",
        "title": "Viking-App ApiREST",
        "contact": {
            "name": "mirazopablo",
            "email": "mirazopablo@gmail.com"
        },
        "version": "1.0"
    },
    "host": "0.0.0.0:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/bookings": {
            "post": {
                "description": "Crea un nuevo turno y reserva el slot en Google Calendar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Crear reserva de turno",
                "operationId": "createBooking",
                "parameters": [
                    {
                        "description": "Booking Payload",
                        "name": "booking",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BookingCreateDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.BookingResponseDto"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/v1/bookings/availability": {
            "get": {
                "description": "Retorna los bloques horarios (Time Slots) disponibles para una fecha específica. Para 'general' devuelve 14 slots; para el resto, 4 slots.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Obtener slots disponibles",
                "operationId": "getAvailability",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fecha en formato YYYY-MM-DD",
                        "name": "date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tipo de equipo/servicio (general | pc | laptop | mobile | gaming)",
                        "name": "deviceType",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AvailabilityResponseDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/v1/bookings/date/{date}": {
            "get": {
                "description": "Obtiene la lista de turnos (bookings) para una fecha específica.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Obtener turnos por fecha",
                "operationId": "getBookingsByDate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fecha en formato YYYY-MM-DD",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Booking"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/v1/bookings/{id}/bot-status": {
            "patch": {
                "description": "Habilita o deshabilita la atención automática por bot (n8n/Evolution API)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Actualizar estado del bot para Handoff",
                "operationId": "updateBotStatus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del turno (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Estado del bot (botActive)",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateBotStatusDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/api/v1/bookings/{id}/status": {
            "patch": {
                "description": "Actualiza el estado del turno tanto en DB como en Google Calendar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bookings"
                ],
                "summary": "Actualizar estado del turno",
                "operationId": "updateBookingStatus",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID del turno (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nuevo estado",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateBookingStatusDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AvailabilityResponseDto": {
            "type": "object",
            "properties": {
                "availableSlots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.TimeSlotDto"
                    }
                },
                "date": {
                    "type": "string"
                }
            }
        },
        "models.Booking": {
            "type": "object",
            "properties": {
                "botActive": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "deviceType": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "googleEventId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timeSlotId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.BookingCreateDto": {
            "type": "object",
            "required": [
                "date",
                "deviceType",
                "fullName",
                "phone",
                "timeSlotId"
            ],
            "properties": {
                "date": {
                    "type": "string"
                },
                "deviceType": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "timeSlotId": {
                    "type": "string"
                }
            }
        },
        "models.BookingResponseDto": {
            "type": "object",
            "properties": {
                "bookingId": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "models.TimeSlotDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "isAvailable": {
                    "type": "boolean"
                },
                "time": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateBookingStatusDto": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
        "models.UpdateBotStatusDto": {
            "type": "object",
            "required": [
                "botActive"
            ],
            "properties": {
                "botActive": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "bearer-jwt": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}